+++ database/migrations/2021_01_23_044558_create_admins_table.php
@@ -16,7 +16,7 @@ class CreateAdminsTable extends Migration
         Schema::create('admins', function (Blueprint $table) {
             $table->id();
             $table->integer("user_id");
-            $table->integer("nik");
+            $table->bigInteger("nik");
             $table->timestamps();
         });
     }
